feat : 신고하기 api 붙이기 #63

+++ app/src/main/java/com/example/fit_i/data/service/CommunalService.kt
@@ -27,7 +27,7 @@ interface CommunalService {
 
     //리뷰 조회
     @GET("api/communal/review/{trainerIdx}")
-    fun getReviewList(@Path("trainerIdx") trainerIdx : Int) : Call<GetReviewListResponse>
+    fun getReviewList(@Path("trainerIdx") trainerIdx: Long) : Call<GetReviewListResponse>
 
     //마이페이지 조회
     @GET("api/communal/mypage")
